feat(view): 興味を尋ねるページをネイティブ対応
